requirements.txt
asgiref==3.8.1
bidict==0.23.1
blinker==1.9.0
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
ConfigArgParse==1.7.1
cssselect2==0.8.0
dj-database-url==3.0.1
Django==5.2
django-cors-headers==4.7.0
django-unfold==0.65.0
djangorestframework==3.16.0
Flask==3.1.1
flask-cors==6.0.1
Flask-Login==0.6.3
fonttools==4.57.0
gevent==25.5.1
geventhttpclient==2.3.4
greenlet==3.2.4
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
locust==2.38.0
locust-cloud==1.26.3
MarkupSafe==3.0.2
msgpack==1.1.1
pillow==11.2.1
platformdirs==4.3.8
psutil==7.0.0
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
python-engineio==4.12.2
python-socketio==5.13.0
pyzmq==27.0.1
reportlab==4.4.0
requests==2.32.3
setuptools==80.9.0
simple-websocket==1.1.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
urllib3==2.4.0
weasyprint==65.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
zope.event==5.1.1
zope.interface==7.2
zopfli==0.2.3.post1
